modify chain id's type

+++ internal/authentication/auth.go
@@ -9,7 +9,7 @@ import (
 func LoadAuthRouter(r *gin.RouterGroup) {
 	r.POST("/login", LoginHandler)
 	r.GET("/identity", VerifyIdentityHandler, func(c *gin.Context) {
-		c.JSON(200, fmt.Sprintf("address:%s  chainid:%d\n", c.GetString("address"), c.GetInt64("chainid")))
+		c.JSON(200, fmt.Sprintf("address:%s  chainid:%d\n", c.GetString("address"), c.GetInt("chainid")))
 	})
 }
 
@@ -32,7 +32,7 @@ func LoginHandler(c *gin.Context) {
 		chainID = 985
 	}
 
-	_, err := Login(address, token, int64(chainID), int64(timestamp), signature)
+	_, err := Login(address, token, int(chainID), int64(timestamp), signature)
 	if err != nil {
 		c.JSON(401, gin.H{"error": err.Error()})
 		return
@@ -58,12 +58,12 @@ func VerifyIdentityHandler(c *gin.Context) {
 		chainID = 985
 	}
 
-	address, err := VerifyIdentity(token, int64(chainID), int64(requestID), signature)
+	address, err := VerifyIdentity(token, int(chainID), int64(requestID), signature)
 	if err != nil {
 		c.AbortWithStatusJSON(401, gin.H{"error": err.Error()})
 		return
 	}
 
 	c.Set("address", address)
-	c.Set("chainid", int64(chainID))
+	c.Set("chainid", int(chainID))
 }
